refactor, feat: feed crud 모듈 이름 변경, 로그인 모듈 연동
